Refactor: remove utils/constants as public API
