+++ src/test/engine-stubs.test.ts
@@ -3,8 +3,9 @@ import { AddressData } from '../key-derivation';
 import { TreeBalance, OutputType, CommitmentType } from '../models';
 import { getTokenDataERC20, TransactNote } from '../note';
 import { Prover } from '../prover/prover';
-import { ZERO_32_BYTE_VALUE, ByteUtils, ByteLength } from '../utils';
+import { ByteUtils, ByteLength } from '../utils';
 import { AbstractWallet, RailgunWallet } from '../wallet';
+import { ZERO_32_BYTE_VALUE } from '../utils/constants';
 
 let balancesStub: SinonStub;
 let treeBalancesStub: SinonStub;
