+++ src/transaction/transaction-batch.ts
@@ -31,7 +31,7 @@ import { POI } from '../poi';
 import { PoseidonMerkleVerifier } from '../abi/typechain';
 import { Memo } from '../note/memo';
 import WalletInfo from '../wallet/wallet-info';
-import { ZERO_ADDRESS } from '../utils';
+import { ZERO_ADDRESS } from '../utils/constants';
 
 export const GAS_ESTIMATE_VARIANCE_DUMMY_TO_ACTUAL_TRANSACTION = 9000;
